Start by simplifying TextureWriter code
