+++ include/ghoul/io/texture/texturewriter.h
@@ -27,7 +27,6 @@
 #define __GHOUL___TEXTUREWRITER___H__
 
 #include <ghoul/misc/exception.h>
-#include <memory>
 #include <string>
 #include <vector>
 
@@ -35,22 +34,45 @@ namespace ghoul::opengl { class Texture; }
 
 namespace ghoul::io {
 
-class TextureWriterBase;
-
 /**
- * This class manages multiple TextureWriterBase and makes them available through one
- * method saveTexture. TextureWriterBases are added through the method addWriter. The
- * class provides a static member, but also allows users to create local variants.
- * TextureWriterBases can be reused between multiple TextureWriters.
+ * A utility class for writing OpenGL textures to image files on disk using the STB
+ * image writer library. This class provides a centralized interface for saving textures
+ * in various common image formats (JPEG, PNG, BMP, TGA).
+ *
+ * The class uses the singleton pattern through its static `ref()` method, providing a
+ * globally accessible instance. The appropriate image format is automatically determined
+ * from the file extension provided in the filename.
+ *
+ * Supported file formats include:
+ * - JPEG (.jpeg, .jpg)
+ * - PNG (.png)
+ * - BMP (.bmp)
+ * - TGA (.tga)
+ *
+ * \note 3D textures are not supported and will result in an error.
+ * \note The class automatically creates the target directory if it doesn't exist.
  */
 class TextureWriter {
 public:
     /**
-     * Exception that gets thrown when there is no writer for the provided \p extension.
+     * Exception that gets thrown when the provided \p extension is not supported.
      */
     struct MissingWriterException final : public RuntimeError {
         explicit MissingWriterException(std::string extension);
         const std::string fileExtension;
+    };
+
+    /**
+     * The exception that gets thrown if there was an error writing the Texture.
+     */
+    struct TextureWriteException final : public RuntimeError {
+        explicit TextureWriteException(std::string name, std::string msg);
+
+        /// The filename that caused the exception to be thrown
+        const std::string filename;
+
+        /// The error message that occurred
+        const std::string errorMessage;
     };
 
     /**
@@ -61,45 +83,32 @@ public:
     static TextureWriter& ref();
 
     /**
-     * Saves the provided \p texture into the \p filename on disk. The correct
-     * TextureWriterBase is determined by the extension of the \p filename.
+     * Saves the provided \p texture into the \p filename on disk. The image file format
+     * is determined by the extension of the \p filename.
      *
      * \param texture The Texture that is to be written to disk
-     * \param filename The target filename for \p filename. The extension of the
-     *        \p filename determines the TextureWriterBase
+     * \param filename The target filename for \p filename.
      *
      * \throw TextureWriteException If there was an error writing the \p filename
-     * \throw MissingWriterException If there was no writer for the specified \p filename
+     * \throw MissingWriterException If the extension in the \p filename is not supported
      * \pre \p filename must not be empty
      * \pre \p filename must have an extension
-     * \pre At least one TextureWriterBase must have been added to the TextureWriter
-     *      before (addWriter)
      */
     void saveTexture(const opengl::Texture& texture, const std::string& filename);
 
     /**
-     * Adds the \p writer to this TextureWriter and makes it available through subsequent
-     * calls to saveTexture. If an extension is supported by multiple TextureWriterBases,
-     * the TextureWriterBase that was added first will be used.
+     * Returns Whether the provided file \p extension is supported by this TextureWriter.
      *
-     * \param writer The writer that is to be added to this TextureWriter
-     *
-     * \pre \p writer must not have been added to this TextureWriter before
+     * \return True if the provided \p extension is supported, false otherwise
      */
-    void addWriter(std::unique_ptr<TextureWriterBase> writer);
+    bool isSupportedExtension(const std::string& extension) const;
 
-private:
     /**
-     * Returns the TextureWriterBase that is responsible for the provided extension.
+     * Returns the supported file extensions.
      *
-     * \param extension The extension for which the TextureWriterBase should be returned
-     * \return The first TextureWriterBase that can write the provided \p extension, or
-     *         `nullptr` if no such file exists
+     * \return The supported file extensions
      */
-    TextureWriterBase* writerForExtension(const std::string& extension);
-
-    /// The list of all registered writers
-    std::vector<std::unique_ptr<TextureWriterBase>> _writers;
+    std::vector<std::string> supportedExtensions() const;
 };
 
 } // namespace ghoul::io
